Move type impls out of values mod
Before this commit `values/mod.rs` is 1700 lines, close to recommended
file length limit.
